Add files via upload
I  analyzed the variance of birth rates in the countries we operate in to drive policy changes. Work through the dataset provided to see when we expect to insure the most births.
I Plotted the total number of births over time using the Month Year field.
By default, Power BI will visualize this using the built-in Date Hierarchy; I changed this to the Date Value instead.
Filtered the page to only show births in France in the last ten years of available data and found which month had the highest birth rate. the second analysis uses the influenza data set to analyze the occurrence of influenza cases each year to drive vaccination policies and identify the trend. I wrote The Dax to create a new calculated column called "Outbreak Check" that uses the CONTAINSSTRING() function, which marks whether there was an influenza outbreak or not for each row of the data.
The TITLE field contains information on the types of influenza events that occurred, and that is why I  checked if the field has the text "Outbreak".I only want to total the influenza cases when an outbreak occurs and use our newly created Outbreak Check field.

I Plotted the total number of influenza cases over time by the kind of outbreak. care about when the case starts, not how long it lasts.

Used a clustered column chart with the SDATE field on the x-axis (Ensure that you are using the date value and not the date hierarchy)

Filtered the visual to focus on France over the last five years of available data. Note that France only has data until 2015.
after all that analysis we can see Which type of variation we see in this influenza dataset, and when we have a spike in outbreak cases. In that case, we have
Cyclical variation in February that variation means data will have a similar increase or decrease at the same point each year and tends to be over a shorter period.